+++ modules/project/ui/src/sap/commerce/toolset/project/view/HybrisProjectView.kt
@@ -75,7 +75,7 @@ open class HybrisProjectView(val project: Project) : TreeStructureProvider, Dumb
     override fun modify(
         parent: AbstractTreeNode<*>,
         children: MutableCollection<AbstractTreeNode<*>>,
-        settings: ViewSettings
+        settings: ViewSettings?
     ): Collection<AbstractTreeNode<*>> {
         if (project.isNotHybrisProject) return children
 
@@ -205,8 +205,8 @@ open class HybrisProjectView(val project: Project) : TreeStructureProvider, Dumb
             ?: true
     }
 
-    private fun isCompactEmptyMiddleFoldersEnabled(settings: ViewSettings) = applicationSettings.hideEmptyMiddleFolders
-            && settings.isHideEmptyMiddlePackages
+    private fun isCompactEmptyMiddleFoldersEnabled(settings: ViewSettings?) = applicationSettings.hideEmptyMiddleFolders
+            && settings?.isHideEmptyMiddlePackages == true
 
     private fun modifyExternalLibrariesNodes(
         children: Collection<AbstractTreeNode<*>>